+++ src/components/github/repo-analysis/RepoResults.tsx
@@ -18,7 +18,7 @@ import {
 	Group as TeamIcon,
 	Assessment as SummaryIcon,
 } from "@mui/icons-material";
-import type { RepoResultsProps } from "./types";
+import type { RepoResultsProps } from "../types/types.ts";
 import TabPanel from "./TabPanel";
 import SummaryTab from "../analysis/tabs/SummaryTab.tsx";
 import CommitsTab from "../analysis/tabs/CommitsTab.tsx";
